rename: Move WarningBottomSheet to common
