+++ src/components/common/bottomsheet/WarningBottomSheet.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import BottomSheet from "../common/bottomsheet/BottomSheet";
-import Button from "../common/atoms/Button";
+import BottomSheet from "./BottomSheet";
+import Button from "../atoms/Button";
 
 export default function WarningBottomSheet({ message, onClose }) {
   return (
